+++ tests/test_resolution.py
@@ -117,7 +117,7 @@ def test_get_resolved_dependencies_with_tilde_requirement_using_json_api():
     assert plist == [
         "pkg:pypi/click@8.1.3",
         "pkg:pypi/flask@2.1.3",
-        "pkg:pypi/importlib-metadata@4.12.0",
+        "pkg:pypi/importlib-metadata@5.0.0",
         "pkg:pypi/itsdangerous@2.1.2",
         "pkg:pypi/jinja2@3.1.2",
         "pkg:pypi/markupsafe@2.1.1",
@@ -147,7 +147,7 @@ def test_without_supported_wheels():
         "pkg:pypi/hyperlink@21.0.0",
         "pkg:pypi/idna@3.4",
         "pkg:pypi/pycparser@2.21",
-        "pkg:pypi/setuptools@65.4.0",
+        "pkg:pypi/setuptools@65.4.1",
         "pkg:pypi/txaio@22.2.1",
     ]